1st version of unoptimized trial division algorithm
This is a smarter trial division algorithm than usually implemented, it
checks at each step whether it should continue instead of blindly
checking every number between 2 and sqrt(n). It can still be optimized
more by using lower-level Rmpz_* functions instead of the operator
overloading of % and /. This is the fastest factoring algorithm to use
if the input number has a factor less than roughly 1000.